requirements.txt
pandas==1.3.3
scikit-learn==0.24.2
numpy==1.19.5
tensorflow
dvc 
notebook 
matplotlib
seaborn 
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib 
types-PyYAML
scipy
Flask 
Flask-Cors
imblearn
-e .    